Controller,error handler and constants added
Added controller, remove all the logic from the contactRoutes and move it to the controller, added errorHandler, that uses constants for all the different types of messages and status codes. Imported the errorHandler and the json parser in the server.js file. Installed AsyncHandler, to handle the functions without the need to use try catch with async only.