Access Token, user id's
Created userModel, that have uname,email and password, validation added. Installed bcrypt and used to hash the password for the user. Imported jsonWebToken which helps for the auth part, that is using the auth token. Every part of the CRUD operations is made private, the way you can access it is with that access token. User controller, have it's own validation part of, registering, logging and getting the information about the current user.